+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [Dashboard improvements #712](https://github.com/farmOS/farmOS/pull/712)
+- [Condense metrics UI #711](https://github.com/farmOS/farmOS/pull/711)
 - [Misc quick form code and documentation improvements #703](https://github.com/farmOS/farmOS/pull/703)
 
 ### Deprecated